feat(trace): frame tracer for HTTP/2, named enums
Closes #610 for the HTTP/2 stack. There was no way to watch what the
client did on the wire, so a bug report could carry a description of
the symptom and nothing else. For H2/H1 the answer was tcpdump plus a
way to decrypt TLS; for QUIC there is not even that, because the TLS
1.3 handshake is ours and exposes no keylog knob.

The stack was opaque at three levels and this fixes two of them, in
the order the issue stages: name the enums, then put the seam where
the frames already funnel. H1 and H3/QUIC are separate work.

WHAT THE SEAM IS. frame.Tracer takes a *FrameInfo — a value struct, so
adding a field later is not an interface break, and so a consumer who
wants qlog or JSON writes a tracer instead of parsing text. It carries
the header plus whichever decoded scalar the frame type has: the
RST_STREAM and GOAWAY codes, GOAWAY's last-stream-id, WINDOW_UPDATE
increments, PUSH_PROMISE's promised id, PING's payload, SETTINGS
parameters. Those are exactly the fields the two bugs the issue cites
turned on -- the flow-control deadlock behind #471/#472 is a
WINDOW_UPDATE that does not appear, and #570 is a GOAWAY code that got
lost.

WHERE IT FIRES. Read side: once in ReadFrame, BEFORE the §6.10
continuity check and before dispatch. That ordering is the point. It
puts two classes of frame in the trace that no Handler can ever see --
an unknown extension type, which §5.5 obliges the codec to drop
silently, and the frame that trips a continuity teardown. Both are
invisible from above and both are what a bug report needs.

Write side: writeHeader is the funnel every Write* method passes
through, save one -- WriteHeaders' coalescing fast path, which writes
its own prefix and so carries its own emit. Without that second site
the most common frame the client sends would be the one frame missing
from every trace. Decoded detail is staged onto the Framer by the
Write method and cleared by the emit, since the funnel does not know
which frame it is writing. writeFrame clears the staging when it
refuses an oversized frame: that is the one rejection reachable AFTER
staging, so without it a frame that never reached the wire would leave
its error code sitting there for the next frame to claim.

ZERO COST WHEN OFF IS A GATE, NOT AN ASPIRATION. A nil Tracer is one
nil compare -- the existing frame benchmarks are unchanged at 0 B/op.
Non-nil is also zero, which is the harder half: an event built on the
stack and handed to an interface method escapes. So the two event
structs live on the heap-resident Framer and are passed by pointer.
Two of them, not one, because conn drives ReadFrame on its reader
goroutine while another goroutine writes under wmu -- a shared scratch
would race, and the reader also calls Write* RE-ENTRANTLY from inside
its own dispatch, which -race cannot catch. Both are pinned:
BenchmarkFramer_*_Traced on the CI bench-gate, and
TestFramer_Trace_ReentrantWriteFromDispatch for the same-goroutine
case. The Framer doc now states the read-half/write-half split conn
has always relied on and that was written down nowhere.

The built-in tracer holds the same line: trace.TextTracer renders into
a reused line buffer, and TestTextTracer_DoesNotAllocatePerFrame
guards it. That test is named to match ci.yml's -run alternation and
./trace is added to that step's package list, because a //go:build
!race gate in a package CI does not name never runs at all -- the
failure mode that comment block already records three times.

SECRETS. Header names, header values and body bytes are absent by
construction: the frame layer sees an opaque HPACK block, and
FrameInfo.Payload is receive-side only, never rendered unless asked,
and POSEIDON_DEBUG=all deliberately does not ask. `payload` has to be
typed out. A debug log is the thing people paste into a public issue.

A RUNTIME SWITCH, NOT A BUILD TAG. -tags poseidondebug already exists
and cannot answer the question that matters, which is a binary someone
already shipped misbehaving in an environment you did not build.
trace.FromEnv reads POSEIDON_DEBUG; loadgen wires it up. `streams` and
`flow` parse and are documented as emitting nothing yet, and Pending()
exists so a program can say so; an unrecognised category is an error,
because a typo in a debug switch that silently turns nothing on costs
an hour of concluding the bug does not reproduce under tracing.

TextTracer.Tracer() looks redundant and is not: a nil *TextTracer
assigned into a frame.Tracer field is a NON-nil interface, so every
emit site's nil check would pass and a build that asked for no tracing
would pay an interface call per frame.

NAMING THE ENUMS pays for itself outside the tracer. frame.FrameType,
Flags, ErrCode, SettingID and FrameHeader now render as their RFC 7540
registry names, so conn.ConnError, conn.StreamError, conn.GoAwayError
and client.StreamResetError say code=FLOW_CONTROL_ERROR where they
said code=3. Flags takes the frame type, because 0x1 is END_STREAM on
DATA and ACK on SETTINGS and a renderer that ignores that shows a
SETTINGS ACK as a protocol violation that never happened. http3 and
quic get functions rather than methods: their constants are bare
uint64, and naming that type would break every caller holding one.

The rendering change was audited across the whole tree with a
type-checking pass over every package with tests, under every build
tag: 114 sites change and all but four are failure-diagnostic messages.
No golden file, no Example output block and no equality assertion on a
formatted string is affected. One site did need fixing --
settings_capacity_test.go used %#x on a SettingID, and fmt routes x/X/q
through Stringer just as it does v/s, so it would have hex-encoded the
NAME and made the diagnostic unreadable at the moment it is needed.

SCOPE. ClientOptions.Tracer reaches every transport that dials a
*conn.Conn. The HTTP/1.1 and HTTP/3 transports accept it and ignore it,
documented as such rather than rejected, so the same configuration
starts working when those seams land.

Also refreshes docs/COVERAGE.md, whose table had been missing http3,
quic, qpack, grpc, header and internal/bufx since they shipped.

Verified: go test -race ./... green, golangci-lint v2.5 clean,
bench-gate green over all seven packages, coverage-gate green with
./trace at 89.9%.

+++ frame/settings_capacity_test.go
@@ -49,8 +49,13 @@ func TestWriteSettings_ScratchFitsAFullTable(t *testing.T) {
 	gotID := SettingID(last[0])<<8 | SettingID(last[1])
 	gotVal := uint32(last[2])<<24 | uint32(last[3])<<16 | uint32(last[4])<<8 | uint32(last[5])
 	if gotID != wantID || gotVal != wantVal {
+		// uint16 conversions, not bare %#x on the SettingID. fmt routes x/X/q
+		// through a Stringer just as it does v/s, so once SettingID names
+		// itself, %#x on one hex-encodes the NAME — id 0x5 prints as
+		// 0x53455454494e47535f... and the diagnostic becomes unreadable at
+		// exactly the moment it is needed.
 		t.Errorf("last entry on the wire = id %#x value %d, want id %#x value %d",
-			gotID, gotVal, wantID, wantVal)
+			uint16(gotID), gotVal, uint16(wantID), wantVal)
 	}
 }
